moved 1d to 2d regridding from cassegrain ray tracing to algorithms.py
