+++ src/astrohack/cassegrain_ray_tracing.py
@@ -1,10 +1,9 @@
 import toolviper
 import xarray as xr
 
-from astrohack.antenna import get_proper_telescope
 from astrohack.utils.validation import custom_unit_checker, custom_plots_checker
 from astrohack.core.cassegrain_ray_tracing import *
-from astrohack.utils import convert_unit, clight, add_caller_and_version_to_dict
+from astrohack.utils import convert_unit, clight, add_caller_and_version_to_dict, regrid_data_onto_2d_grid
 from astrohack.utils.phase_fitting import aips_like_phase_fitting
 from astrohack.visualization.plot_tools import create_figure_and_axes, close_figure
 from typing import Union
